requirements.txt
httpx[http2]>=0.28.1
python-telegram-bot>=22.5
